split "import-path" into two rules "mechanics" and "circular-deps"
