+++ stack/eslint-plugin-packlets/src/index.ts
@@ -2,7 +2,8 @@
 // See LICENSE in the project root for license information.
 
 import { TSESLint } from '@typescript-eslint/experimental-utils';
-import { importPath } from './import-path';
+import { importPath } from './mechanics';
+import { circularDeps } from './circular-deps';
 
 interface IPlugin {
   rules: { [ruleName: string]: TSESLint.RuleModule<string, unknown[]> };
@@ -10,8 +11,10 @@ interface IPlugin {
 
 const plugin: IPlugin = {
   rules: {
-    // Full name: "@rushstack/packlets/import-path"
-    'import-path': importPath
+    // Full name: "@rushstack/packlets/mechanics"
+    'import-path': importPath,
+    // Full name: "@rushstack/packlets/circular-deps"
+    'circular-deps': circularDeps
   }
 };
