+++ stack/eslint-plugin-packlets/src/mechanics.ts
